+++ CHANGES.md
@@ -3,6 +3,7 @@
 - [NEW] Added `DesignDocumentManager.list()` to return all design documents defined in a database.
 - [NEW] Added an optional `SettableViewParameters.STALE_NO` constant for the default omitted case of
   the stale parameter on a view request.
+- [NEW] Added `descending` option for changes feed.
 
 # 2.4.3 (2016-05-05)
 - [IMPROVED] Reduced the length of the User-Agent header string.